requirementsPip.txt
brotlipy==0.7.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.4
cloudpickle==2.0.0
cryptography==36.0.0
cycler==0.11.0
Cython==0.29.28
cytoolz==0.11.2
dask==2022.2.1
decorator==5.1.1
docutils==0.18.1
fsspec==2022.2.0
future==0.18.2
idna==3.3
imagecodecs-lite==2019.12.3
imageio==2.6.1
imutils==0.5.3
Kivy==1.11.1
Kivy-Garden==0.1.4
kiwisolver==1.3.2
locket==0.2.0
matplotlib==3.2.2
mkl-fft==1.3.0
mkl-random==1.1.1
mkl-service==2.3.0
networkx==2.3
numpy==1.18.5
OpenPIV==0.23.4
packaging==21.3
partd==1.2.0
Pillow==9.0.1
pip==21.2.2
pycparser==2.21
Pygments==2.11.2
pyOpenSSL==22.0.0
pyparsing==3.0.4
PySocks==1.7.1
python-dateutil==2.8.2
PyWavelets==1.3.0
PyYAML==6.0
requests==2.27.1
scikit-image==0.19.2
scipy==1.5.0
setuptools==58.0.4
sip==4.19.13
six==1.16.0
tifffile==2019.7.26.2
toolz==0.11.2
tornado==6.1
urllib3==1.26.8
wheel==0.37.1
win-inet-pton==1.1.0
wincertstore==0.2
